# arcx_backend/openapi.yaml
openapi: 3.0.3
info:
  title: ARCX Platform API
  version: 1.0.0
  description: |-
    ## ARCX — Algorithmic Reserve Currency Exchange

    ARCX is an INR-denominated, NAV-backed digital asset platform. Investors deposit INR which is converted to ARCX tokens at the current NAV price. The vault holds a diversified portfolio (SPY, TLT, GLD) rebalanced algorithmically.

    ### Authentication
    All protected endpoints require a **Bearer JWT** in the `Authorization` header:
    ```
    Authorization: Bearer <access_token>
    ```

    Obtain tokens via `POST /api/auth/token/` or `POST /api/v1/auth/register`.

    ### Idempotency
    Financial mutation endpoints (deposit, withdraw, transfer) require an `Idempotency-Key` UUID header to prevent duplicate transactions.

    ### Error Format
    All errors follow: `{"error": "message", "code": "SNAKE_CASE_CODE"}`
  contact:
    name: ARCX Engineering
    email: engineering@arcx.in
  license:
    name: Proprietary
paths:
...
  /api/v1/kyc/submit:
    post:
      operationId: kyc_submit
      description: |-
        Submits a KYC document reference for verification. ARCX stores only the external reference ID returned by the KYC provider (DigiLocker, Onfido, etc.) — never the raw document itself.

        **Tier limits:**
        - `tier_1` (Aadhaar/DL) → ₹10,000/day
        - `tier_2` (PAN/Passport) → ₹1,00,000/day
        - `tier_3` (Full address proof) → Unlimited

        **Demo:** Pass any string as `document_ref` (e.g. `DEMO_REF_001`) to test the flow without a real KYC provider.
      summary: Submit KYC document for verification
      tags:
      - KYC
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KYCSubmitRequest'
            examples:
              Tier1AadhaarSubmission:
                value:
                  tier: tier_1
                  document_type: aadhaar
                  document_ref: DIGILOCKER_REF_ABC123
                summary: Tier 1 Aadhaar submission
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/KYCSubmitRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/KYCSubmitRequest'
        required: true
      security:
      - jwtAuth: []
      - BearerAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KYCSubmitResponse'
              examples:
                KYCApproved:
                  value:
                    kyc_id: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                    tier: tier_1
                    status: approved
                    verified_at: '2025-06-01T15:30:00Z'
                    message: KYC tier_1 approved. You can now transact up to Rs.10,000/day.
                  summary: KYC approved
          description: ''
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KYCErrorResponse'
              examples:
                AlreadyApproved:
                  value:
                    error: KYC is already approved.
                    code: KYC_ALREADY_APPROVED
                  summary: Already approved
          description: KYC already approved for this user
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KYCErrorResponse'
          description: Validation error — invalid tier/document_type combination
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KYCErrorResponse'
          description: JWT missing or expired
  /api/v1/kyc/status:
    get:
      operationId: kyc_status
      description: Returns the user's current overall KYC status, the highest approved
        tier, the resulting daily transaction limit, and the full list of all KYC
        submissions.
      summary: Get current KYC status and history
      tags:
      - KYC
      security:
      - jwtAuth: []
      - BearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KYCStatusResponse'
              examples:
                KYCStatusResponse:
                  value:
                    kyc_status: approved
                    highest_tier: tier_1
                    daily_limit_inr: Rs.10,000
                    records:
                    - id: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                      tier: tier_1
                      status: approved
                      document_type: aadhaar
                      verified_at: '2025-06-01T15:30:00Z'
                      submitted_at: '2025-06-01T15:29:50Z'
                  summary: KYC status response
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KYCErrorResponse'
          description: JWT missing or expired
...
components:
  schemas:
...
    DocumentTypeEnum:
      enum:
      - aadhaar
      - pan
      - passport
      - dl
      type: string
      description: |-
        * `aadhaar` - Aadhaar Card
        * `pan` - PAN Card
        * `passport` - Passport
        * `dl` - Driving License
...
    KYCErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
      required:
      - code
      - error
    KYCStatusResponse:
      type: object
      properties:
        kyc_status:
          type: string
          description: Overall KYC status on the user account
        highest_tier:
          type: string
          nullable: true
          description: Highest approved KYC tier
        daily_limit_inr:
          type: string
          description: Current daily transaction limit based on KYC tier
        records:
          type: array
          items:
            type: object
            additionalProperties: {}
            description: 'KYC record: id, tier, status, document_type, verified_at,
              submitted_at'
          description: All KYC submissions for this user
      required:
      - daily_limit_inr
      - highest_tier
      - kyc_status
      - records
    KYCSubmitRequest:
      type: object
      description: |-
        POST /api/v1/kyc/submit

        document_ref: The external ID returned by your KYC provider
                      (DigiLocker, Onfido, CKYC, etc.).
                      We never store the actual document — only this reference.
      properties:
        tier:
          $ref: '#/components/schemas/TierEnum'
        document_type:
          $ref: '#/components/schemas/DocumentTypeEnum'
        document_ref:
          type: string
          minLength: 1
          description: External reference ID from KYC provider. Never send raw document
            data.
          maxLength: 255
      required:
      - document_ref
      - document_type
      - tier
    KYCSubmitResponse:
      type: object
      properties:
        kyc_id:
          type: string
          format: uuid
          description: Unique ID of the KYC record
        tier:
          type: string
          description: 'KYC tier that was approved: tier_1, tier_2, or tier_3'
        status:
          type: string
          description: 'Status of this KYC record: approved, pending, or rejected'
        verified_at:
          type: string
          format: date-time
          nullable: true
        message:
          type: string
          description: Human-readable summary of the approval with daily limit
      required:
      - kyc_id
      - message
      - status
      - tier
      - verified_at
...
    TierEnum:
      enum:
      - tier_1
      - tier_2
      - tier_3
      type: string
      description: |-
        * `tier_1` - Tier 1 — Aadhaar OTP
        * `tier_2` - Tier 2 — PAN + Selfie
        * `tier_3` - Tier 3 — Full Address Proof
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...